requirements/dev.txt
-r base.txt
pytest>=7.0
pytest-sphinx>=0.6.0
pytest-cov>=4.0
beautifulsoup4>=4.11.0
black>=23.0
